Move disassembly logic from cpu.rs to disassembly.rs
The disassemble, fetch_instruction, build_instruction_repr, and
fetch_operand methods didn't use any CPU state - they only needed
memory access and the static disassembly tables. Moving them to
disassembly.rs makes that module self-contained and reduces cpu.rs
by ~90 lines.
