+++ core/src/disassembly.rs
@@ -1,3 +1,5 @@
+use crate::memory::Memory;
+
 use AddressingMode::*;
 
 #[derive(Debug)]
@@ -3096,3 +3098,95 @@ pub static PREFIXED_DISASSEMBLY: [DisassemblyInfo; 256] = [
         addressing_mode: Implied,
     },
 ];
+
+pub fn disassemble(memory: &Memory, start_addr: u16, end_addr: u16) -> Vec<(u16, String)> {
+    let mut current_addr = start_addr;
+    let mut instruction_reprs = Vec::new();
+
+    loop {
+        if current_addr > end_addr {
+            break;
+        };
+
+        let instruction = fetch_instruction(memory, current_addr);
+        let num_bytes = instruction.num_bytes as u32;
+
+        if current_addr as u32 + num_bytes > end_addr as u32 {
+            break;
+        };
+
+        let operand_start_addr = current_addr + 1 + (instruction.prefixed as u16);
+        let repr = build_instruction_repr(memory, operand_start_addr, instruction);
+        instruction_reprs.push((current_addr, repr));
+        current_addr += num_bytes as u16;
+    }
+
+    instruction_reprs
+}
+
+fn fetch_instruction(memory: &Memory, addr: u16) -> &'static DisassemblyInfo<'static> {
+    let opcode = memory.cpu_read(addr);
+
+    if opcode == 0xCB {
+        let next_byte = memory.cpu_read(addr + 1);
+        &PREFIXED_DISASSEMBLY[next_byte as usize]
+    } else {
+        &DISASSEMBLY[opcode as usize]
+    }
+}
+
+fn build_instruction_repr(memory: &Memory, addr: u16, instruction: &DisassemblyInfo) -> String {
+    let (operand, _) = fetch_operand(memory, addr, instruction);
+    let mut repr = String::from(instruction.repr);
+
+    match &instruction.addressing_mode {
+        AddressingMode::Implied => (),
+        AddressingMode::Imm8 => {
+            repr.find("d").map(|index| {
+                repr.replace_range(index..index + 2, &format!("{:02x}", operand));
+            });
+        }
+        AddressingMode::Imm16 => {
+            repr.find("d").map(|index| {
+                repr.replace_range(index..index + 3, &format!("{:04x}", operand));
+            });
+        }
+        AddressingMode::Addr16 => {
+            repr.find("a").map(|index| {
+                repr.replace_range(index..index + 3, &format!("{:04x}", operand));
+            });
+        }
+        AddressingMode::ZeroPageOffset => {
+            repr.find("a").map(|index| {
+                repr.replace_range(index..index + 2, &format!("{:04x}", operand));
+            });
+        }
+        AddressingMode::SignedAddrOffset => {
+            repr.find("r").map(|index| {
+                repr.replace_range(index..index + 2, &format!("{:04x}", operand));
+            });
+        }
+    };
+
+    repr
+}
+
+fn fetch_operand(memory: &Memory, addr: u16, instruction: &DisassemblyInfo) -> (u16, u32) {
+    // (operand, bytes_read)
+    let lsb = memory.cpu_read(addr) as u16;
+    match &instruction.addressing_mode {
+        AddressingMode::Implied => (0, 0),
+        AddressingMode::Imm8 => (lsb, 0),
+        AddressingMode::Imm16 | AddressingMode::Addr16 => {
+            let msb = memory.cpu_read(addr + 1) as u16;
+            (msb << 8 | lsb, 2)
+        }
+        AddressingMode::ZeroPageOffset => (0xFF00 | lsb, 1),
+        AddressingMode::SignedAddrOffset => {
+            let lsb = lsb as i8;
+            // The offset assumes that the PC has already been incremented
+            let target_addr = ((addr + 1) as i32) + (lsb as i32);
+            (target_addr as u16, 1)
+        }
+    }
+}
